requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.9.0
django-sendgrid-v5==1.2.2
djangorestframework==3.14.0
mysqlclient==2.1.1
PyJWT==2.6.0
python-http-client==3.3.7
pytz==2022.7.1
sendgrid==6.9.7
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tzdata==2022.7
